Refactoring for readability/ease of maintainance - moved influxdb classes into separate packages. Moved constants to constants package. Cleaned up imports.
