+++ graphite_influxdb/__init__.py
@@ -1,1 +1,2 @@
-from .classes import *
+from .classes.finder import InfluxdbFinder
+from .classes.reader import InfluxdbReader
